pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["ManØfCulture <rrs.discord@gmail.com>"]
description = "Atla Discord Bot. Made By ManØfCulture#6666An all-purpose bot for your Discord servers ! Check out What I do here : https://top.gg/bot/767624993507901470 Invite me to your servers ! : https://dsc.gg/atla-bot"
name = "atla"
version = "1.6.0"
[tool.poetry.dependencies]
pip = "20.0.2"
PyNaCl = "^1.4.0"
dblpy = "^0.4.0"
discord = "^1.0.1"
discord-py-slash-command = "^1.2.2"
flask = "^2.0.1"
python = "^3.8"
python-dotenv = "^0.15.0"
pytz = "^2021.1"
youtube-dl = "^2021.06.06"
replit = "^3.0.0"
praw = "^7.3.0"
discord-components = "^1.1.4"
topggpy = "^1.3.0"
[tool.poetry.dev-dependencies]
